light-rag-agent/LightRAG/requirements.light.txt
fastapi==0.110.0
uvicorn==0.35.0
python-dotenv==1.1.0
httpx==0.28.1
openai==1.70.0
lightrag-hku==1.3.0
nano-vectordb==0.0.4.3
pydantic==2.11.1
typing_extensions==4.13.0
anyio==4.9.0
tiktoken==0.9.0
pandas==2.2.2
pipmaster==0.4.0
pydantic-ai==0.0.19
